+++ cocos/3d/physics/cocos/builtin-body.ts
@@ -3,8 +3,8 @@ import { clamp } from '../../../core/vmath';
 import { intersect } from '../../geom-utils';
 import { ICollisionCallback, ICollisionEvent, ICollisionType as ICollisionEventType, PhysicsWorldBase, RigidBodyBase } from '../api';
 import { ERigidBodyType } from '../physic-enum';
-import { BuiltInShape } from './built-in-shape';
-import { BuiltInWorld } from './built-in-world';
+import { BuiltInShape } from './builtin-shape';
+import { BuiltInWorld } from './builtin-world';
 
 /**
  * Built-in static collider, no physical forces involved
